Fixes to make the source build

+++ lustre/tests/badarea_io.c
@@ -34,7 +34,7 @@
 
 int main(int argc, char **argv)
 {
-	int rc;
+	long long rc;
 	int fd;
 	void *volatile buf = (void *)0x4096000;
 	void *volatile fd_ptr;
